docs: Clarify CPU is default wildcard in Expert Mode
Update the "Bytes" mode documentation in the README to specify that the CPU acts as the default wildcard device.

This change clarifies that if no `*` is explicitly used in the device allocation string, the remainder of the model will be automatically assigned to the CPU. This helps users better understand the default behavior and prevent confusion.